Move NavButton and HelperTextRow as well
