+++ src/interface/guide/components/CastDetail.tsx
@@ -6,6 +6,8 @@ import { useState, useMemo, useCallback, useRef } from 'react';
 import { TipBox } from './TipBox';
 import GuideDataWrapper, {
   HelperText,
+  HelperTextRow,
+  NavButton,
   FilterBadge,
   PerfBadgeCount,
   PerfBadgeDivider,
@@ -468,46 +470,10 @@ const AdditionalContentHeading = styled.div`
   letter-spacing: 0.5px;
 `;
 
-// ── Shared primitives (exported for use by CastSummary, CastOverview, BuffUptimeBar, CastSequence) ──
-
-export const HelperTextRow = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin-top: 4px;
-`;
-
-export const NavButton = styled.button<{ disabled?: boolean }>`
-  min-width: 32px;
-  height: 32px;
-  cursor: ${(props) => (props.disabled ? 'not-allowed' : 'pointer')};
-  background: rgba(0, 0, 0, 0.25);
-  border: 1px solid rgba(255, 255, 255, ${(props) => (props.disabled ? '0.04' : '0.1')});
-  border-radius: 6px;
-  color: ${(props) => (props.disabled ? 'rgba(255,255,255,0.15)' : '#fab700')};
-  font-size: 1.8rem;
-  font-weight: 400;
-  transition: all 0.15s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0 8px;
-  line-height: 1;
-  -webkit-tap-highlight-color: transparent;
-
-  &:hover:not(:disabled) {
-    background: rgba(250, 183, 0, 0.12);
-    border-color: rgba(250, 183, 0, 0.35);
-  }
-
-  &:active:not(:disabled) {
-    transform: scale(0.9);
-    background: rgba(250, 183, 0, 0.18);
-  }
-`;
-
-// StatCard, StatsGrid, PerfBadge*, FilterBadge are defined in GuideDataWrapper and re-exported from there.
+// HelperTextRow, NavButton, StatCard, StatsGrid, PerfBadge*, FilterBadge are defined in GuideDataWrapper and re-exported from there.
 export {
+  HelperTextRow,
+  NavButton,
   StatCard,
   StatCardValue,
   StatCardDivider,
